requirements.txt
# VibeOS - Production Dependencies
# Core AI/ML
langgraph==0.2.16
langchain==0.2.14
langchain-groq==0.1.9
langchain-community==0.2.12
anthropic==0.34.2
google-generativeai==0.8.0

# APIs & Integrations
tweepy==4.14.0
google-api-python-client==2.147.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
requests==2.32.3
httpx==0.27.2

# Data & Storage
sqlalchemy==2.0.35
pandas==2.2.3
pydantic==2.9.2
python-dotenv==1.0.1

# UI
streamlit==1.39.0
plotly==5.24.1
pillow==10.4.0

# Utilities
beautifulsoup4==4.12.3
lxml==5.3.0
python-dateutil==2.9.0
pytz==2024.2
tenacity==9.0.0

# Development
pytest==8.3.3
black==24.8.0
ruff==0.6.9
